moving hf generators to new module
