+++ examples/ra-dit/ra_dit/evaluation_benchmarks/mmlu.py
@@ -5,7 +5,7 @@ import re
 import pandas as pd
 from pydantic import PrivateAttr
 
-from fed_rag.generators.hf_peft_model import HFPeftModelGenerator
+from fed_rag.generators.huggingface import HFPeftModelGenerator
 from fed_rag.types.rag_system import RAGSystem
 
 from .base import BaseBenchmark, ExamplePred, ScoredExamplePred
