+++ src/fed_rag/generators/huggingface/__init__.py
@@ -1,0 +1,4 @@
+from .hf_peft_model import HFPeftModelGenerator
+from .hf_pretrained_model import HFPretrainedModelGenerator
+
+__all__ = ["HFPeftModelGenerator", "HFPretrainedModelGenerator"]
