+++ tests/generators/test_hf_peft.py
@@ -15,7 +15,7 @@ from transformers import (
 
 from fed_rag.base.generator import BaseGenerator
 from fed_rag.base.tokenizer import EncodeResult
-from fed_rag.generators.hf_peft_model import HFPeftModelGenerator
+from fed_rag.generators.huggingface import HFPeftModelGenerator
 from fed_rag.tokenizers.hf_pretrained_tokenizer import HFPretrainedTokenizer
 
 
@@ -120,7 +120,9 @@ def test_hf_pretrained_generator_class_init_no_load(
     assert generator.tokenizer.unwrapped == tokenizer
 
 
-@patch("fed_rag.generators.hf_peft_model.prepare_model_for_kbit_training")
+@patch(
+    "fed_rag.generators.huggingface.hf_peft_model.prepare_model_for_kbit_training"
+)
 @patch.object(PeftModel, "from_pretrained")
 @patch.object(AutoModelForCausalLM, "from_pretrained")
 @patch.object(AutoTokenizer, "from_pretrained")
@@ -194,7 +196,7 @@ def test_generate() -> None:
     mock_model.generate.assert_called_once()
 
 
-@patch("fed_rag.generators.hf_peft_model.F")
+@patch("fed_rag.generators.huggingface.hf_peft_model.F")
 def test_compute_target_sequence_proba(
     mock_torch_functional: MagicMock,
 ) -> None:
@@ -232,7 +234,7 @@ def test_huggingface_extra_missing() -> None:
     """Test extra is not installed."""
 
     modules = {"transformers": None}
-    module_to_import = "fed_rag.generators.hf_peft_model"
+    module_to_import = "fed_rag.generators.huggingface.hf_peft_model"
 
     if module_to_import in sys.modules:
         original_module = sys.modules.pop(module_to_import)
@@ -246,7 +248,9 @@ def test_huggingface_extra_missing() -> None:
             ValueError,
             match=re.escape(msg),
         ):
-            from fed_rag.generators.hf_peft_model import HFPeftModelGenerator
+            from fed_rag.generators.huggingface.hf_peft_model import (
+                HFPeftModelGenerator,
+            )
 
             HFPeftModelGenerator("fake_name", "fake_base_name")
